Time freezing and blocks
The player can now freeze time. When time is frozen, every rigidbody (except for the player's) will be frozen. When the player unfreezes, the rigidbodies will go back to their original velocity. The block is a good example of this, when you push it then freeze, it'll still move when you unfreeze.